# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 拼单平台接口
  version: 0.2.0
  description: |
    T001：框架元数据与健康检查。
    T002：后台身份权限、商品图片、商品库存管理、小程序商品展示。
    金额一律整数分；数量为十进制字符串（≤3 位小数）；时间 ISO 8601 UTC；
    分页 page 从 1 起，pageSize 默认 10、上限 50；响应包络 {data, requestId}，错误 {code, message, requestId, details?}。
servers:
  - url: http://127.0.0.1:3000
    description: 本地联调
paths:
...
  /api/mini/v1/conversations:
    get:
      operationId: listMiniConversationHistory
      summary: "本人会话历史分页，含结束会话"
      security: [{ bearerAuth: [] }]
      parameters: [{"name":"page","in":"query","schema":{"type":"integer","minimum":1}},{"name":"pageSize","in":"query","schema":{"type":"integer","minimum":1,"maximum":100}}]
      responses:
        '200': {"description":"本人会话历史分页，含结束会话","content":{"application/json":{"schema":{"type":"object","properties":{"data":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/CsConversation"}},"total":{"type":"integer"},"page":{"type":"integer"},"pageSize":{"type":"integer"}}},"requestId":{"type":"string"}}}}}}
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
    post:
      operationId: createMiniConversation
      security: [{ bearerAuth: [] }]
      summary: 发起/恢复会话（已有未结束会话幂等返回；无客服在线=留言语义）
      responses:
        '201':
          description: 会话视图
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    properties:
                      conversation:
                        type: object
                        properties:
                          id: { type: string, format: uuid }
                          status: { type: string, enum: [queued, active, ended, converted] }
                          hasAgent: { type: boolean }
        '401': { $ref: '#/components/responses/Error401' }
  /api/mini/v1/conversations/current:
    get:
      operationId: getCurrentMiniConversation
      security: [{ bearerAuth: [] }]
      summary: 本人未结束会话（无则 conversation=null 空态）
      responses:
        '200':
          description: 会话或空态
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    properties:
                      conversation: { type: [object, 'null'] }
        '401': { $ref: '#/components/responses/Error401' }
  /api/mini/v1/conversations/{id}/messages:
    get:
      operationId: listMiniConversationMessages
      security: [{ bearerAuth: [] }]
      summary: 增量拉取消息（afterSeq 断线补取；≤100 条/页；内部备注不含；他人 404）
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: afterSeq, in: query, schema: { type: integer, minimum: 0, default: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: 增量消息
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    required: [messages, lastSeq, nextSeq]
                    properties:
                      messages:
                        type: array
                        items:
                          type: object
                          required: [seq, sender, kind, content, createdAt]
                          properties:
                            seq: { type: integer }
                            sender: { type: string, enum: [user, agent, system] }
                            kind: { type: string, enum: [text, image, card] }
                            content: { type: object }
                            createdAt: { type: string, format: date-time }
                      lastSeq: { type: integer, description: 会话水位 }
                      nextSeq: { type: integer, description: 本页已返回消息游标 }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
    post:
      operationId: sendMiniConversationMessage
      security: [{ bearerAuth: [] }]
      summary: 发送消息（clientMessageId 幂等；ended/converted 409）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientMessageId, kind, content]
              properties:
                clientMessageId: { type: string, maxLength: 64 }
                kind: { type: string, enum: [text, image, card] }
                content:
                  type: object
                  properties:
                    text: { type: string, maxLength: 2000 }
                    mediaAssetId: { type: string, format: uuid }
                    cardKind: { type: string, enum: [product, order, group, refund] }
                    cardRefId: { type: string, format: uuid }
      responses:
        '200':
          description: 幂等返回原消息（duplicated=true）
        '201':
          description: 已发送（seq 单调）
          content:
            application/json:
              schema:
                type: object
                required: [data, requestId]
                properties:
                  requestId: { type: string }
                  data:
                    type: object
                    properties:
                      message:
                        type: object
                        required: [seq, sender, kind, content, createdAt]
                        properties:
                          seq: { type: integer }
                          sender: { type: string }
                          kind: { type: string }
                          content: { type: object }
                          createdAt: { type: string, format: date-time }
                      duplicated: { type: boolean }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
        '409': { $ref: '#/components/responses/Error409' }
  /api/mini/v1/conversations/{id}/end:
    post:
      operationId: endMiniConversation
      security: [{ bearerAuth: [] }]
      summary: 用户结束会话（幂等）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: 已结束 }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/mini/v1/tickets:
    post:
      operationId: createMiniTicket
      security: [{ bearerAuth: [] }]
      summary: 用户提交售后工单
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, title, description]
              properties:
                type: { type: string, enum: [group_issue, payment_issue, refund_issue, product_issue, shipment_issue, complaint, other] }
                title: { type: string, maxLength: 60 }
                description: { type: string, maxLength: 2000 }
                orderId: { type: string, format: uuid }
                clientTicketId: { type: string, minLength: 1, maxLength: 64, description: 本用户范围幂等键 }
      responses:
        '201': { description: 已创建（open） }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
    get:
      operationId: listMiniTickets
      security: [{ bearerAuth: [] }]
      summary: 本人工单列表
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 10 } }
      responses:
        '200': { description: 分页工单（仅本人） }
        '401': { $ref: '#/components/responses/Error401' }
  /api/mini/v1/tickets/{id}:
    get:
      operationId: getMiniTicket
      security: [{ bearerAuth: [] }]
      summary: 本人工单详情（他人 404）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: 工单详情 }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
  /api/mini/v1/tickets/{id}/feedback:
    post:
      operationId: feedbackMiniTicket
      security: [{ bearerAuth: [] }]
      summary: 用户反馈（确认解决/继续追问；本人）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [satisfied]
              properties:
                clientActionId: { type: string, minLength: 1, maxLength: 64, description: 工单与发送者范围幂等键 }
                satisfied: { type: boolean }
                text: { type: string, maxLength: 2000 }
      responses:
        '200': { description: 已记录；同 clientActionId 幂等，异内容409 }
        '401': { $ref: '#/components/responses/Error401' }
        '404': { $ref: '#/components/responses/Error404' }
  /api/admin/v1/cs/queue:
    get:
      operationId: listCsQueue
      security: [{ bearerAuth: [] }]
      summary: 排队中会话（agent:manage）
      responses:
        '200': { description: 排队列表 }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
  /api/admin/v1/cs/heartbeat:
    post:
      operationId: csHeartbeat
      security: [{ bearerAuth: [] }]
      summary: 客服在线心跳（agent:manage）
      responses:
        '200': { description: ok }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
  /api/admin/v1/cs/conversations:
    get:
      operationId: listCsConversations
      security: [{ bearerAuth: [] }]
      summary: 我的会话分页（agent:manage）
      parameters: [{ name: status, in: query, schema: { type: string, enum: [queued, active, ended, converted] } }]
      responses:
        '200': { description: 会话列表 }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
  /api/admin/v1/cs/conversations/{id}/accept:
    post:
      operationId: acceptCsConversation
      security: [{ bearerAuth: [] }]
      summary: 接入（queued→active；并发恰一人成功，其余 409）
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        '200': { description: 已接入 }
        '401': { $ref: '#/components/responses/Error401' }
        '403': { $ref: '#/components/responses/Error403' }
        '409': { $ref: '#/components/responses/Error409' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 管理员会话 token（admin 域）
    miniUserAuth:
      type: http
      scheme: bearer
      description: 小程序用户会话 token（user 域，与 admin 不可互用）
  responses:
...
    Error400:
      description: 参数校验失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error401:
      description: 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Error404:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Error409:
      description: 状态冲突 / 不可上架（details 为原因清单）/ 图片被引用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Error403:
      description: 无权限
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }
        details:
          type: array
          items: { type: string }
...
    CsConversation: {"type":"object","properties":{"id":{"type":"string","format":"uuid"},"status":{"type":"string","enum":["queued","active","ended","converted"]},"hasAgent":{"type":"boolean"}}}
...